GitHub action to push tag based on VERSION
Uses anothrNick/github-tag-action but the logic
is much simpler: whenever the `VERSION` file is
pushed to the `master` branch, use that file's content
as the version number, prefix it with `v`, and push
a tag with that name to the remote. The remote
could be xspec/xspec or a fork of it.